Move xvideos to app module
