+++ app/build.gradle
@@ -38,7 +38,6 @@ dependencies {
     implementation project(":decoders:animesproject")
     implementation project(":decoders:anitubex")
     implementation project(":decoders:onepiecex")
-    implementation project(":decoders:xvideos")
 
     kapt "com.android.databinding:compiler:$plugin_version"
 
@@ -48,6 +47,9 @@ dependencies {
     kapt 'com.brunodles:environmentmods-processor:0.2.0'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+
+    implementation 'com.google.code.gson:gson:2.8.2'
 
     implementation 'com.android.support:multidex:1.0.3'
     implementation 'com.android.support.constraint:constraint-layout:1.1.2'
