+++ app/src/main/kotlin/brunodles/animewatcher/decoders/UrlChecker.kt
@@ -1,7 +1,6 @@
 package brunodles.animewatcher.decoders
 
 import android.content.Intent
-import brunodles.animacurse.XvideosFactory
 import brunodles.animesorion.AnimesOrionFactory
 import brunodles.animesproject.AnimesProjectFactory
 import brunodles.animewatcher.explorer.Episode
