--- decoders/xvideos/build.gradle
@@ -1,5 +1,0 @@
-apply from: rootProject.file("gradle/decoder_build.gradle")
-
-dependencies {
-    compile 'com.google.code.gson:gson:2.8.2'
-}
