+++ settings.gradle
@@ -6,4 +6,3 @@ include ':decoders:animesorion'
 include ':decoders:animesproject'
 include ':decoders:anitubex'
 include ':decoders:onepiecex'
-include ':decoders:xvideos'
